+++ example/android/app/src/main/kotlin/com/hoc/example/MainActivity.kt
@@ -1,0 +1,8 @@
+package com.hoc.example
+
+import io.flutter.embedding.android.FlutterActivity
+
+/**
+ * MainActivity
+ */
+class MainActivity: FlutterActivity()