+++ lib/rx_shared_preferences.dart
@@ -11,4 +11,3 @@ export 'src/logger/default_logger.dart';
 export 'src/logger/logger.dart';
 export 'src/logger/logger_adapter.dart';
 export 'src/model/key_and_value.dart';
-export 'src/rx_shared_preferences_impl.dart';