+++ test/streams_test.dart
@@ -8,7 +8,7 @@ void main() {
   WidgetsFlutterBinding.ensureInitialized();
 
   group('Test Stream', () {
-    const Map<String, dynamic> kTestValues = <String, dynamic>{
+    const kTestValues = <String, dynamic>{
       'flutter.String': 'hello world',
       'flutter.bool': true,
       'flutter.int': 42,
@@ -21,7 +21,10 @@ void main() {
     setUp(() async {
       SharedPreferences.setMockInitialValues(kTestValues);
 
-      rxPrefs = RxSharedPreferences(await SharedPreferences.getInstance());
+      rxPrefs = RxSharedPreferences(
+        await SharedPreferences.getInstance(),
+        const DefaultLogger(),
+      );
     });
 
     tearDown(() async => await rxPrefs.dispose());
@@ -29,8 +32,7 @@ void main() {
     test(
       'Stream will emit error when read value is not valid type, or emit null when value is not set',
       () async {
-        final Stream<int> intStream =
-            rxPrefs.getIntStream('bool'); // Actual: Stream<bool>
+        final intStream = rxPrefs.getIntStream('bool'); // Actual: Stream<bool>
         await expectLater(
           intStream,
           emitsAnyOf([
@@ -39,7 +41,7 @@ void main() {
           ]),
         );
 
-        final Stream<List<String>> listStringStream =
+        final listStringStream =
             rxPrefs.getStringListStream('String'); // Actual: Stream<String>
         await expectLater(
           listStringStream,
@@ -49,7 +51,7 @@ void main() {
           ]),
         );
 
-        final Stream<int> noSuchStream =
+        final noSuchStream =
             rxPrefs.getIntStream('###String'); // Actual: Stream<String>
 
         await expectLater(
@@ -98,7 +100,7 @@ void main() {
         ///
         /// Bool
         ///
-        final Stream<bool> streamBool = rxPrefs.getBoolStream('bool');
+        final streamBool = rxPrefs.getBoolStream('bool');
         final expectStreamBoolFuture = expectLater(
           streamBool,
           emitsInOrder([anything, false, true, false, true, false]),
@@ -112,7 +114,7 @@ void main() {
         ///
         /// Double
         ///
-        final Stream<double> streamDouble = rxPrefs.getDoubleStream('double');
+        final streamDouble = rxPrefs.getDoubleStream('double');
         final expectStreamDoubleFuture = expectLater(
           streamDouble,
           emitsInOrder([anything, 0.3333, 1, 2, isNull, 3, isNull, 4]),
@@ -128,7 +130,7 @@ void main() {
         ///
         /// Int
         ///
-        final Stream<int> streamInt = rxPrefs.getIntStream('int');
+        final streamInt = rxPrefs.getIntStream('int');
         final expectStreamIntFuture = expectLater(
           streamInt,
           emitsInOrder([anything, 1, isNull, 2, 3, isNull, 3, 2, 1]),
@@ -145,7 +147,7 @@ void main() {
         ///
         /// String
         ///
-        final Stream<String> streamString = rxPrefs.getStringStream('String');
+        final streamString = rxPrefs.getStringStream('String');
         final expectStreamStringFuture = expectLater(
           streamString,
           emitsInOrder([anything, 'h', 'e', 'l', 'l', 'o', isNull]),
@@ -160,8 +162,7 @@ void main() {
         ///
         /// List<String>
         ///
-        final Stream<List<String>> streamListString =
-            rxPrefs.getStringListStream('List');
+        final streamListString = rxPrefs.getStringListStream('List');
         final expectStreamListStringFuture = expectLater(
           streamListString,
           emitsInOrder([
@@ -216,7 +217,7 @@ void main() {
       for (final v in expected.skip(1)) {
         await rxPrefs.setStringList(
           'List',
-          <String>['before', 'dispose'],
+          v,
         );
         await Future.delayed(Duration.zero);
       }
@@ -244,7 +245,7 @@ void main() {
     });
 
     test('Emit null when clearing', () async {
-      final Stream<List<String>> stream = rxPrefs.getStringListStream('List');
+      final stream = rxPrefs.getStringListStream('List');
 
       final later = expectLater(
         stream,
@@ -352,91 +353,11 @@ void main() {
       expect(identical(rxPrefs3, rxPrefs1) && rxPrefs3 != null, isFalse);
     });
 
-    test('Stream is broadcast', () {
+    test('Stream is single-subscription stream', () {
       final stream = rxPrefs.getStringListStream('List');
-      expect(stream.isBroadcast, isTrue);
+      expect(stream.isBroadcast, isFalse);
       stream.listen(null);
-      stream.listen(null);
-      expect(true, true);
-    });
-
-    test(
-      'Stream replay last data event when subscribing',
-      () async {
-        const values = [
-          ['1', '1', '1'],
-          ['2', '2', '2'],
-          ['3', '3', '3'],
-          ['4', '4', '4'],
-          ['5', '5', '5'],
-          ['6', '6', '6'],
-          ['7', '7', '7'],
-        ];
-        final stream = rxPrefs.getStringListStream('List');
-
-        // emits initial value and all values.
-        final expect1 = <dynamic>[anything, ...values];
-        var index1 = 0;
-        stream.listen(expectAsync1(
-          (event) => expect(event, expect1[index1++]),
-          count: expect1.length,
-        ));
-
-        await rxPrefs.setStringList('List', values[0]);
-        await Future.delayed(Duration.zero);
-
-        // replay values[0] and emits rest.
-        const expect2 = values;
-        var index2 = 0;
-        stream.listen(expectAsync1(
-          (event) => expect(event, expect2[index2++]),
-          count: expect2.length,
-        ));
-
-        for (final v in values.skip(1)) {
-          await rxPrefs.setStringList('List', v);
-          await Future.delayed(Duration.zero);
-        }
-      },
-    );
-
-    test('Stream replay last error event when subscribing', () async {
-      // replay error event
-      final listStringsStream = rxPrefs.getStringListStream('int');
-      listStringsStream.listen(
-        null,
-        onError: expectAsync1(
-          (error) => expect(error, isA<TypeError>()),
-        ),
-      );
-
-      await Future.delayed(const Duration(milliseconds: 100));
-      listStringsStream.listen(
-        null,
-        onError: expectAsync1(
-          (error) => expect(error, isA<TypeError>()),
-        ),
-      );
-
-      // replay error event and continue emits events after subscribing.
-      final intStream = rxPrefs.getIntStream('String');
-      intStream.listen(
-        expectAsync1((event) => expect(event, 42)),
-        onError: expectAsync1(
-          (error) => expect(error, isA<TypeError>()),
-        ),
-      );
-
-      await Future.delayed(const Duration(milliseconds: 100));
-      intStream.listen(
-        expectAsync1((event) => expect(event, 42)),
-        onError: expectAsync1(
-          (error) => expect(error, isA<TypeError>()),
-        ),
-      );
-
-      await Future.delayed(const Duration(milliseconds: 100));
-      await rxPrefs.setInt('String', 42);
+      expect(() => stream.listen(null), throwsStateError);
     });
   });
 }