+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.3.0 - May 28, 2020
+
+*   **Breaking change**: returned stream is a ***single-subscription*** stream.
+*   Internal implementation refactor.
+
 ## 1.2.0 - Apr 20, 2020
 
 *   Breaking change: support for `rxdart` 0.24.x.